Removed nested rows
Removes the construction of nested rows (rows in rows), when a literal
stream contains a sequence of literals that has the same name.

This feature adds more confusion to the output, due to the fact that
a nested row is a csv row on its own. A removal of this feature
keeps the implementation simple and predictable.